% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_cov_matrix.R
\name{compute_cov_matrix}
\alias{compute_cov_matrix}
\title{Compute covariance matrix for a Gaussian process}
\usage{
compute_cov_matrix(points, cov_func, ...)
}
\arguments{
\item{points}{A data frame with \code{n} rows and \code{d} columns.}

\item{cov_func}{A covariance model. Options are \code{white_noise}, \code{brownian_motion}, \code{exponential}, and \code{matern_model}. For all functions (except for \code{brownian_motion}), one can provide \code{sig2}.
Also, for the \code{exponential} model, one has to provide the parameter \code{beta}, and for the \code{matern_model} model, the parameters \code{nu} and \code{beta}.
See \url{https://avramaral.github.io/STAT294/rpack.html} for more details.}

\item{...}{Additional parameters. See above.}
}
\value{
An \code{n} by \code{n} covariance matrix.
}
\description{
Compute covariance matrix for a Gaussian process
}
\examples{
x <- seq(1, 10, 1)
y <- seq(1, 10, 1)
z <- expand.grid(x, y)
compute_cov_matrix(points = z, cov_func = "matern_model", sig2 = 1, nu = 1, beta = 10)
}
