% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPcov.R
\docType{package}
\name{GPcov}
\alias{GPcov}
\title{GPcov: A package for computing the covariance matrix for stationary Gaussian processes}
\description{
The \code{GPcov} package provides one important function, namely, \code{compute_cov_matrix()}.
}
\section{\code{GPcov} functions}{

The \code{compute_cov_matrix()} function computes the covariance matrix for a given model.
}

